fix: seperate db.tsx file and supabase.tsx file
